refactor: change repository name common/dto to response
최상단 공통 부모인 `support` 와 하위 레포지토리에서의 이름이 충돌되는 것으로 생각되어 이름 변경
