+++ src/main/kotlin/com/kdongsu5509/user/adapter/in/web/friends/FriendsRequestCommandController.kt
@@ -1,6 +1,6 @@
 package com.kdongsu5509.user.adapter.`in`.web.friends
 
-import com.kdongsu5509.support.common.dto.APIResponse
+import com.kdongsu5509.support.response.APIResponse
 import com.kdongsu5509.user.adapter.`in`.web.friends.dto.CreateFriendRequest
 import com.kdongsu5509.user.adapter.`in`.web.friends.dto.CreateFriendRequestResponse
 import com.kdongsu5509.user.adapter.`in`.web.friends.dto.FriendRelationshipResponse
