+++ src/main/kotlin/com/kdongsu5509/user/adapter/in/web/friends/FriendsRequestReadController.kt
@@ -1,6 +1,6 @@
 package com.kdongsu5509.user.adapter.`in`.web.friends
 
-import com.kdongsu5509.support.common.dto.APIResponse
+import com.kdongsu5509.support.response.APIResponse
 import com.kdongsu5509.user.adapter.`in`.web.friends.dto.ReceivedFriendRequestResponse
 import com.kdongsu5509.user.adapter.`in`.web.friends.dto.ReceivedFriendRequestResponseDetail
 import com.kdongsu5509.user.application.port.`in`.friend.ReadFriendRequestUseCase
